Refactor: Move external integrations to an `integrations` module
This pull request refactors the codebase to group external integrations into a new `integrations` module. This improves code organization and makes it easier to manage integrations with external services.

**Changes:**

- Created a new `lua/core/integrations` directory.
- Moved `lazygit.lua`, `fzf.lua`, and `usql.lua` from `lua/core` to `lua/core/integrations`.
- Updated `init.lua` to require the moved modules from their new location.
- Updated `lua/core/integrations/fzf.lua` to use the new path.
- Removed `fzf_grep`, `Fzfgrep` and `fuzzy_find_func` from `lua/core/search.lua` to avoid duplication.

This refactoring improves the project structure and makes it more scalable for future integrations.
