+++ init.lua
@@ -40,10 +40,10 @@ run_if("search", function()
 	require("core.search")
 end)
 run_if("lazygit", function()
-	require("core.lazygit")
+	require("core.integrations.lazygit")
 end)
 run_if("fzf", function()
-	require("core.fzf")
+	require("core.integrations.fzf")
 end)
 run_if("statusline", function()
 	require("core.statusline")
@@ -52,7 +52,7 @@ run_if("colorscheme", function()
 	require("core.colorscheme").setup()
 end)
 run_if("usql", function()
-	require("core.usql")
+	require("core.integrations.usql")
 end)
 
 -- Load optional post_init.lua file
